Move Folder and FolderLoader test to separate test module
